requirements.txt
requests
pandas
nltk
python-dotenv